backend: check stock while checkouting

+++ backend/src/main/java/com/bugenzhao/bookstore_backend/entity/db/Book.java
@@ -40,7 +40,7 @@ public class Book {
 
     @NotNull
     @Min(value = 0)
-    private Integer inventory;
+    private Long inventory;
 
     @NotNull
     private String image;
